+++ test/boarding_util_test.dart
@@ -3,27 +3,46 @@ import 'package:boarding/util.dart';
 
 testUtil() {
   group('Testing utilities', () {
-    test('Random number', () {
-      var rn = randomNum(10);
-      expect(rn, lessThan(10));
-    });
     test('Random integer', () {
-      var ri = randomInt(10);
-      expect(ri, lessThan(10));
+      var r = randomInt(10);
+      expect(r, lessThan(10));
+    });
+    test('Random double', () {
+      var r = randomDouble(1.9);
+      expect(r, lessThan(1.9));
+    });
+    test('Random number', () {
+      var r = randomNum(10.1);
+      expect(r, lessThan(10.1));
+    });
+    test('Random range number', () {
+      var r = randomRangeNum(0.8, 9.9);
+      expect(r, inClosedOpenRange(0.8, 9.9));
+    });
+    test('Random sign number', () {
+      var r = randomSignNum(10);
+      expect(r, inClosedOpenRange(-10, 10));
     });
     test('Random list element', () {
       var list = [22, 45.8, 33.8, 44, 8];
-      var re = randomElement(list);
-      expect(list.contains(re), isTrue);
+      var r = randomElement(list);
+      expect(list.contains(r), isTrue);
     });
     test('Random color', () {
-      var rc = randomColor();
-      expect(colorList().contains(rc), isTrue);
+      var r = randomColor();
+      expect(colorList().contains(r), isTrue);
     });
     test('Random color code', () {
-      var rcc = randomColorCode();
-      print(rcc);
-      expect(colorMap().containsValue(rcc), isTrue);
+      var r = randomColorCode();
+      expect(colorMap().containsValue(r), isTrue);
+    });
+    test('Random bool', () {
+      var r = randomBool();
+      expect(r, anyOf([isTrue, isFalse]));
+    });
+    test('Random sign', () {
+      var r = randomSign(10);
+      expect(r, anyOf([-1, 1]));
     });
   });
 }