+++ CHANGELOG.md
@@ -11,6 +11,8 @@ based on [Semantic Versioning](http://semver.org/)
 + replace FallingPiece with MovablePiece in Board of the drop example
 + show Direction in import in the crash and drop examples
 + add 'black' and 'white' to color list in util
++ rename randomNumber to randomSignNum in util
++ add more random tests
 
 *0.2.1* 2015-03-08